Move checkpoint funtions from runner to a new sub-package
