+++ docs/en/api/checkpoint.rst
@@ -1,0 +1,34 @@
+.. role:: hidden
+    :class: hidden-section
+
+mmengine.checkpoint
+===================================
+
+.. contents:: mmengine.checkpoint
+   :depth: 2
+   :local:
+   :backlinks: top
+
+.. currentmodule:: mmengine.checkpoint
+
+.. autosummary::
+   :toctree: generated
+   :nosignatures:
+   :template: classtemplate.rst
+
+   CheckpointLoader
+
+.. autosummary::
+   :toctree: generated
+   :nosignatures:
+
+   load_checkpoint
+   save_checkpoint
+   load_state_dict
+   get_state_dict
+   weights_to_cpu
+   find_latest_checkpoint
+   get_deprecated_model_names
+   get_external_models
+   get_mmcls_models
+   get_torchvision_models
