+++ docs/zh_cn/index.rst
@@ -103,6 +103,7 @@
    mmengine.dataset <api/dataset>
    mmengine.infer <api/infer>
    mmengine.device <api/device>
+   mmengine.checkpoint <api/checkpoint>
    mmengine.hub <api/hub>
    mmengine.logging <api/logging>
    mmengine.visualization <api/visualization>
