+++ mmengine/testing/__init__.py
@@ -2,11 +2,13 @@
 from .compare import (assert_allclose, assert_attrs_equal,
                       assert_dict_contains_subset, assert_dict_has_keys,
                       assert_is_norm_layer, assert_keys_equal,
-                      assert_params_all_zeros, check_python_script)
+                      assert_params_all_zeros, assert_tensor_equal,
+                      check_python_script)
 from .runner_test_case import RunnerTestCase
 
 __all__ = [
-    'assert_allclose', 'assert_dict_contains_subset', 'assert_keys_equal',
-    'assert_attrs_equal', 'assert_dict_has_keys', 'assert_is_norm_layer',
-    'assert_params_all_zeros', 'check_python_script', 'RunnerTestCase'
+    'assert_allclose', 'assert_tensor_equal', 'assert_dict_contains_subset',
+    'assert_keys_equal', 'assert_attrs_equal', 'assert_dict_has_keys',
+    'assert_is_norm_layer', 'assert_params_all_zeros', 'check_python_script',
+    'RunnerTestCase'
 ]
